requirements.txt
comtypes==1.2.0
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
Pillow==9.4.0
PyAudio==0.2.14
pycaw==20230407
